cleanups and split out patches sub-mod
